+++ kwsCheckVariables.cxx
@@ -67,7 +67,7 @@ bool Parser::CheckVariables(const char* regEx)
   // Construct the list of variables to check
   typedef std::pair<std::string,int> PairType;
   std::vector<PairType> ivars;
-  long int pos = 0;
+  size_t pos = 0;
   while(pos != -1)
     {
     std::string var = this->FindVariable(buffer,pos+1,buffer.size(),pos);
@@ -177,9 +177,9 @@ bool Parser::CheckVariables(const char* regEx)
 }
 
 /** Find the first ivar in the source code */
-std::string Parser::FindVariable(std::string & buffer, long int start, long int end,long int & pos)
+std::string Parser::FindVariable(std::string & buffer, size_t start, size_t end,size_t & pos)
 {
-  long int posSemicolon = buffer.find(";",start);
+  size_t posSemicolon = buffer.find(";",start);
   while(posSemicolon != -1 && posSemicolon<end)
     {
     // We try to find the word before that
